backend/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
alembic==1.5.4
aniso8601==6.0.0
appdirs==1.4.4
astroid==2.2.5
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
awscli==1.19.27
Babel==2.9.0
beautifulsoup4==4.9.3
botocore==1.20.27
bs4==0.0.1
chardet==4.0.0
Click==7.0
colorama==0.4.3
distlib==0.3.1
docutils==0.15.2
ecdsa==0.13.2
filelock==3.0.12
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.17.1
gunicorn==20.0.4
idna==3.1
iniconfig==1.1.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.0
loader==2017.9.11
Logbook==1.5.3
logcc==2018.3.1
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
packaging==20.9
pendulum==2.1.2
pluggy==0.13.1
pony==0.7.14
postgres==3.0.0
psycopg2-binary==2.8.2
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.6.0
python-dotenv==0.17.0
python-editor==1.0.4
python-jose==3.2.0
python-jose-cryptodome==1.3.2
pytz==2019.1
pytzdata==2020.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.3.4
six==1.12.0
soupsieve==2.2.1
SQLAlchemy==1.3.4
termcc==2019.8.2
terminaltables==3.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
websockets==8.1
Werkzeug==1.0.1
wrap==2018.3.1
wrapt==1.11.1
WTForms==2.3.3
yarl==1.6.3
